extract login logic to useLogin hook & rewrite test
