+++ src/hooks/__mocks__/useLogin.js
@@ -1,0 +1,5 @@
+export const mockLogin = jest.fn();
+const mock = jest.fn().mockImplementation(() => {
+  return [false, mockLogin];
+});
+export default mock;
